+++ src/clojure_lsp/feature/code_lens.clj
@@ -24,12 +24,12 @@
 
 (defn ^:private var-definitions-lens [filename kondo-config analysis]
   (->> (q/find-var-definitions analysis filename true)
-       (remove (partial f.diagnostic/exclude-public-var? kondo-config))))
+       (remove (partial f.diagnostic/exclude-public-definition? kondo-config))))
 
 (defn ^:private keyword-definitions-lens
   [filename kondo-config analysis]
   (->> (q/find-keyword-definitions analysis filename)
-       (remove (partial f.diagnostic/exclude-public-var? kondo-config))))
+       (remove (partial f.diagnostic/exclude-public-definition? kondo-config))))
 
 (defn reference-code-lens [uri db]
   (let [analysis (:analysis @db)