+++ src/pod/clojure_lsp/api.clj
@@ -75,7 +75,8 @@
     (.toString baos "utf-8")))
 
 (def api-vars
-  {'clojure-lsp.api/analyze-project! api/analyze-project!
+  {'clojure-lsp.api/analyze-project-and-deps! api/analyze-project-and-deps!
+   'clojure-lsp.api/analyze-project-only! api/analyze-project-only!
    'clojure-lsp.api/clean-ns! api/clean-ns!
    'clojure-lsp.api/diagnostics api/diagnostics
    'clojure-lsp.api/format! api/format!